+++ misc/changelog.md
@@ -10,6 +10,7 @@
 
 - Limit max threads to avoid system instability.
 - Update ExifTool to `13.16` (Windows).
+- Rework smallest lossless mode.
 
 ### Fixed